% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn.measures}
\alias{fn.measures}
\title{Create a List of Evaluation Metrics}
\usage{
fn.measures()
}
\value{
A data frame with the following columns:
  \itemize{
    \item \code{names}: A character vector of metric names.
    \item \code{type}: An integer vector representing the evaluation type of each metric.
  }
  The total number of metrics equals the length of the \code{names} vector.
}
\description{
This function creates a data frame containing the names of evaluation metrics,
their corresponding types, and the total count of metrics.
Each metric is categorized by its type, which indicates whether a higher
or lower value represents better performance.
}
\details{
The \code{type} vector categorizes metrics as follows:
\itemize{
  \item \code{1}: Metrics where a higher value indicates better performance.
  \item \code{0}: Metrics where a lower value indicates better performance.
}

Below is a summary of the metrics and their evaluation types:

\tabular{lll}{
\strong{Metric Name} \tab \strong{Type} \tab \strong{Interpretation} \cr
accuracy           \tab 1 \tab higher is better \cr
auprc_macro        \tab 1 \tab higher is better \cr
auprc_micro        \tab 1 \tab higher is better \cr
auprc_samples      \tab 1 \tab higher is better \cr
auprc_weighted     \tab 1 \tab higher is better \cr
average_precision  \tab 1 \tab higher is better \cr
clp                \tab 0 \tab lower is better \cr
coverage           \tab 0 \tab lower is better \cr
f1                 \tab 1 \tab higher is better \cr
hamming_loss       \tab 0 \tab lower is better \cr
macro_auc          \tab 1 \tab higher is better \cr
macro_F1           \tab 1 \tab higher is better \cr
macro_precision    \tab 1 \tab higher is better \cr
macro_recall       \tab 1 \tab higher is better \cr
margin_loss        \tab 0 \tab lower is better \cr
micro_auc          \tab 1 \tab higher is better \cr
micro_F1           \tab 1 \tab higher is better \cr
micro_precision    \tab 1 \tab higher is better \cr
micro_recall       \tab 1 \tab higher is better \cr
mlp                \tab 0 \tab lower is better \cr
one_error          \tab 0 \tab lower is better \cr
precision          \tab 1 \tab higher is better \cr
ranking_loss       \tab 0 \tab lower is better \cr
recall             \tab 1 \tab higher is better \cr
roc_auc_macro      \tab 1 \tab higher is better \cr
roc_auc_micro      \tab 1 \tab higher is better \cr
roc_auc_samples    \tab 1 \tab higher is better \cr
roc_auc_weighted   \tab 1 \tab higher is better \cr
subset_accuracy    \tab 1 \tab higher is better \cr
wlp                \tab 0 \tab lower is better \cr
}
}
\examples{
# Create the data frame of metrics
metric_df <- fn.measures()

# Display the resulting data frame
print(metric_df)

}
