% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friedman-nemenyi.R
\name{generate.boxplots.1}
\alias{generate.boxplots.1}
\title{Generate Boxplots of Method Performance}
\usage{
generate.boxplots.1(data, methods, save.dir, measure.name, width, height)
}
\arguments{
\item{data}{A data frame containing performance metrics for different methods.}

\item{methods}{A character vector specifying the names of the methods (columns)
in the data frame.}

\item{save.dir}{A character string specifying the directory where the boxplots 
will be saved.}
}
\value{
A boxplot of the performance of the specified methods.
}
\description{
This function generates boxplots for the performance of various methods
based on the provided data. It saves the plots as PDF files in the specified
directory.
}
\examples{
generate.boxplots.1(data = results_data, methods = c("Method1", "Method2"), save.dir = "output_directory")
}
